Fixed routing for ocs-connected static ring by making OCS transparent when building nbr2if
Pulled apart link initialisation and creation of neighbour table. Links
need to be fully initialized for OCS to be able to report neighbour
relationship. Making OCS transparent means routing tables are built as
if the links are directly connected on a ring - just who is connected to
whom is decided by ocs port map